requirements.txt
Flask==3.1.3
Werkzeug==3.1.7
pandas==3.0.2
python-dotenv==1.2.2
requests==2.33.1
